tracking down why index can corrupt on mono 2.10.1

+++ Firkin.Test/TFirkinDictionary.cs
@@ -47,7 +47,9 @@ namespace Droog.Firkin.Test {
             if(_dictionary != null) {
                 _dictionary.Dispose();
             }
-            Directory.Delete(_path, true);
+            if(Directory.Exists(_path)) {
+                Directory.Delete(_path, true);
+            }
         }
 
         [Test]
@@ -173,6 +175,169 @@ namespace Droog.Firkin.Test {
                 Assert.AreEqual(dictionary.Count, d.Count);
                 foreach(var pair in dictionary) {
                     Assert.AreEqual(pair.Value, d[pair.Key]);
+                }
+            }
+        }
+
+        [Test, Explicit]
+        public void Run_till_crash_with_parallel_merge() {
+            var r = new Random(1234);
+            uint serial = 0;
+            var keys = new Queue<string>();
+            AddItems(keys, 1000);
+            uint last = 0;
+            var mergeCounter = 0;
+            var inMerge = false;
+            using(var d = new FirkinDictionary<string, uint>(_path)) {
+                new Thread(() => {
+                    while(true) {
+                        if(mergeCounter > 100000) {
+                            inMerge = true;
+                            Interlocked.Exchange(ref mergeCounter, 0);
+                            var s = serial;
+                            Console.WriteLine("{0} iterations, merge time", s);
+                            var limit = s - 20000;
+                            var remove = (from entry in d where entry.Value < limit select entry.Key).ToArray();
+                            Console.WriteLine("will remove {0} of {1}", remove.Length, d.Count);
+                            foreach(var key in remove) {
+                                d.Remove(key);
+                            }
+                            var preMerge = d.Count;
+                            d.Merge();
+                            var postMerge = d.Count;
+                            inMerge = false;
+                            Console.WriteLine("pre: {0} / post: {1}", preMerge, postMerge);
+                            foreach(var file in Directory.GetFiles(_path)) {
+                                Console.WriteLine(Path.GetFileName(file));
+                            }
+                        }
+                        Thread.Sleep(1000);
+                    }
+                }) { IsBackground = true }.Start();
+                while(true) {
+                    Interlocked.Increment(ref mergeCounter);
+                    var nextKey = keys.Dequeue();
+                    if(d.ContainsKey(nextKey)) {
+                        last = d[nextKey];
+                    }
+                    d[nextKey] = ++serial;
+                    keys.Enqueue(r.Next(10) == 1 ? nextKey : Guid.NewGuid().ToString());
+                    if(inMerge) {
+                        Thread.Sleep(10);
+                    }
+                }
+            }
+        }
+
+        [Test, Explicit]
+        public void Run_till_crash_with_parallel_writes() {
+            var r = new Random(1234);
+            int serial = 0;
+            var keys = new Queue<string>();
+            AddItems(keys, 1000);
+            uint last = 0;
+            var mergeCounter = 0;
+            var inMerge = false;
+            var faults = new List<Exception>();
+            using(var d = new FirkinDictionary<string, uint>(_path)) {
+                for(var i = 0; i < 10; i++) {
+                    var workerId = i;
+                    var worker = new Thread(() => {
+                        try {
+                            _log.DebugFormat("worker {0} started", workerId);
+                            while(true) {
+                                Interlocked.Increment(ref mergeCounter);
+                                string nextKey;
+                                lock(keys) {
+                                    nextKey = keys.Dequeue();
+                                }
+                                if(d.ContainsKey(nextKey)) {
+                                    last = d[nextKey];
+                                }
+                                var v = (uint)Interlocked.Increment(ref serial);
+                                d[nextKey] = v;
+                                lock(keys) {
+                                    keys.Enqueue(r.Next(10) == 1 ? nextKey : Guid.NewGuid().ToString());
+                                }
+                                if(inMerge) {
+                                    Thread.Sleep(1000);
+                                }
+                            }
+                        } catch(Exception e) {
+                            Console.WriteLine("Worker {0} failed: {1}\r\n{2}", workerId, e.Message, e.StackTrace);
+                            faults.Add(e);
+                        }
+                    }) { IsBackground = true };
+                    worker.Start();
+                }
+                while(true) {
+                    if(faults.Any()) {
+                        throw faults.First();
+                    }
+                    if(mergeCounter > 100000) {
+                        try {
+                            inMerge = true;
+                            Interlocked.Exchange(ref mergeCounter, 0);
+                            var s = serial;
+                            Console.WriteLine("{0} iterations, merge time", s);
+                            var limit = s - 20000;
+                            var remove = (from entry in d where entry.Value < limit select entry.Key).ToArray();
+                            Console.WriteLine("will remove {0} of {1}", remove.Length, d.Count);
+                            foreach(var key in remove) {
+                                d.Remove(key);
+                            }
+                            var preMerge = d.Count;
+                            d.Merge();
+                            var postMerge = d.Count;
+                            inMerge = false;
+                            Console.WriteLine("pre: {0} / post: {1}", preMerge, postMerge);
+                            foreach(var file in Directory.GetFiles(_path)) {
+                                Console.WriteLine(Path.GetFileName(file));
+                            }
+                        } catch(Exception e) {
+                            Console.WriteLine("merger failed: {0}\r\n{1}", e.Message, e.StackTrace);
+                        }
+                    }
+                    Thread.Sleep(1000);
+                }
+            }
+        }
+
+        [Test, Explicit]
+        public void Run_till_crash_with_serial_merge() {
+            var r = new Random(1234);
+            uint serial = 0;
+            var keys = new Queue<string>();
+            AddItems(keys, 1000);
+            uint last = 0;
+            var mergeCounter = 0;
+            using(var d = new FirkinDictionary<string, uint>(_path)) {
+                while(true) {
+                    Interlocked.Increment(ref mergeCounter);
+                    var nextKey = keys.Dequeue();
+                    if(d.ContainsKey(nextKey)) {
+                        last = d[nextKey];
+                    }
+                    d[nextKey] = ++serial;
+                    keys.Enqueue(r.Next(10) == 1 ? nextKey : Guid.NewGuid().ToString());
+                    if(mergeCounter > 100000) {
+                        Interlocked.Exchange(ref mergeCounter, 0);
+                        var s = serial;
+                        Console.WriteLine("{0} iterations, merge time", s);
+                        var limit = s - 20000;
+                        var remove = (from entry in d where entry.Value < limit select entry.Key).ToArray();
+                        Console.WriteLine("will remove {0} of {1}", remove.Length, d.Count);
+                        foreach(var key in remove) {
+                            d.Remove(key);
+                        }
+                        var preMerge = d.Count;
+                        d.Merge();
+                        var postMerge = d.Count;
+                        Console.WriteLine("pre: {0} / post: {1}", preMerge, postMerge);
+                        foreach(var file in Directory.GetFiles(_path)) {
+                            Console.WriteLine(Path.GetFileName(file));
+                        }
+                    }
                 }
             }
         }
